refactor: update UIHandler and file handling in Score class
- Renamed displayGuessSummaryHeader to displayGuessSummarySeparator for clarity and adjusted message formatting.
- Removed the displayGuessSummaryFooter method to streamline the UIHandler class.
- Updated file reading in FileManager to trim lines before adding to the list, ensuring cleaner data handling.
- Enhanced Score class to create score files if they do not exist and improved validation checks for lists and file paths.
