+++ src/code/ca/bcit/comp2522/gameproject/wordGame/FileManager.java
@@ -72,7 +72,7 @@ final class FileManager
 
             while((line = bufferedReader.readLine()) != null)
             {
-                lines.add(line);
+                lines.add(line.trim());
             }
         }
         catch(final IOException error)
